add SpigotAdapter

+++ api/src/main/java/me/hsgamer/minigamemapcore/api/data/RealPosition.java
@@ -11,8 +11,8 @@ public class RealPosition {
     private double x = 0;
     private double y = 0;
     private double z = 0;
-    private double yaw = 0;
-    private double pitch = 0;
+    private float yaw = 0;
+    private float pitch = 0;
 
     public static RealPosition fromString(String string) {
         String[] split = string.split(",", 5);
@@ -20,8 +20,8 @@ public class RealPosition {
         position.setX(Double.parseDouble(split[0].trim()));
         position.setY(Double.parseDouble(split[1].trim()));
         position.setZ(Double.parseDouble(split[2].trim()));
-        position.setYaw(Double.parseDouble(split[3].trim()));
-        position.setPitch(Double.parseDouble(split[4].trim()));
+        position.setYaw(Float.parseFloat(split[3].trim()));
+        position.setPitch(Float.parseFloat(split[4].trim()));
         return position;
     }
